chore(zones): implemented code review suggestions, moved useZonesTableColumns.tsx to subdir
https://github.com/canonical/maas-ui/pull/5588#pullrequestreview-2567045896
