+++ src/app/base/components/GenericTable/_index.scss
@@ -1,4 +1,4 @@
-.generic-table {
+.p-generic-table {
   thead, tbody {
     display: table-row-group;
     width: 100%;
