+++ src/app/images/components/SMImagesTable/_index.scss
@@ -1,4 +1,4 @@
-.generic-table {
+.p-generic-table {
   .status {
     width: 15rem;
   }
